Got the navigation up and running. So when you click on the start quiz. It will run a new activity
However. There was a issue trying to retun to the past_quiz fragment on the nav after starting the activity. I was working on it for a solid hour. The simple solution was just to not allow the user to go back. Therefore fixing the bug. So the only way they can view a past quiz now is by initially clicking on it before they start a new quiz.